feat: add OneBot implementation

+++ Robin.Abstractions/Message/Entities/ImageData.cs
@@ -1,3 +1,3 @@
 namespace Robin.Abstractions.Message.Entities;
 
-public record ImageData(string File, string? Type, string? Url, bool? UseCache, bool? UseProxy, int? Timeout) : SegmentData;
+public record ImageData(string File, string? Type, string? Url, bool? UseCache, bool? UseProxy, double? Timeout) : SegmentData;
